refactor: remove redundant casts in tests (68 CodeQL alerts)
Removes 68 redundant casts flagged by CodeQL across 12 test files:
- 22x cs/useless-cast-to-self (e.g. (uint)expr where expr is already uint)
- 44x cs/useless-upcast on non-null operands (e.g. (T?)rule)
- 2x redundant 'as' upcasts feeding object?-typed reflection helpers

All removed casts are behavior-preserving: they either cast a value to its
own static type, or upcast a non-null operand that flows into a
params object?[] reflection helper where the static type is erased at the
parameter boundary. Method resolution in those helpers is by name + arg
count only, so runtime dispatch is unchanged.

Deliberately NOT touched (left as CodeQL false positives):
- (T?)null casts feeding params object?[] helpers: a bare null collapses to
  the whole params array (TargetParameterCountException), so the cast is
  load-bearing.
- (T?)null inside anonymous-object initializers (CS0828).

Verified: dotnet build (Release) + deterministic test suite green
(3903/3904; the single failure is a pre-existing date/ordering flake in
SupportBundleCoverageSweepTests that passes in isolation).

Part of the QZP drive-to-zero campaign.

+++ tests/SwfocTrainer.Tests/Meg/MegWave3FinalTests.cs
@@ -375,7 +375,7 @@ public sealed class MegWave3FinalTests
         var nameBytes = Encoding.ASCII.GetBytes("test.xml");
         var nameTableSize = (uint)(4 + nameBytes.Length);
         uint nameCount = 1, fileCount = 1;
-        var dataStart = (uint)(24 + nameTableSize + 20);
+        var dataStart = (24 + nameTableSize + 20);
         var totalSize = (int)(dataStart + 10);
         var bytes = new byte[totalSize];
